% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{number_state_map}
\alias{number_state_map}
\title{number_state_map}
\usage{
number_state_map()
}
\value{
a map
}
\description{
Uses the state profile data set to display population date of persons over 
the age of 65 on US Map.
}
