% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{persongs_65_plot}
\alias{persongs_65_plot}
\title{persons_65_p}
\usage{
persongs_65_plot()
}
\value{

}
\description{
Plot population points of Americans over the age of 65 by year and projections
to 2060.
}
