% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{state65plus_m}
\alias{state65plus_m}
\title{state65plus_m}
\usage{
state65plus_m(df)
}
\value{

}
\description{
Uses the state profile data set to display population date of persons over 
the age of 65 on US Map.
}
